fix #417
