+++ app/core/ts/tools/TranslateTool.ts
@@ -1,6 +1,7 @@
 import {GRID_SIZE,
         LEFT_MOUSE_BUTTON,
-        SPACEBAR_KEY} from "core/utils/Constants";
+        SPACEBAR_KEY,
+        WIRE_SNAP_THRESHOLD} from "core/utils/Constants";
 
 import {Vector,V} from "Vector";
 import {Camera} from "math/Camera";
@@ -14,8 +15,33 @@ import {Component} from "core/models/Component";
 import {Action} from "core/actions/Action";
 import {GroupAction} from "core/actions/GroupAction";
 import {CopyGroupAction} from "core/actions/CopyGroupAction";
-import {TranslateAction} from "core/actions/transform/TranslateAction";
 import {CreateGroupPostTranslateAction} from "core/actions/transform/GroupPostTranslateActionFactory";
+import {Wire} from "core/models/Wire";
+
+function SnapPos(obj: Component): void {
+    function DoSnap(wire: Wire, x: number, c: number): number {
+        if (Math.abs(x - c) <= WIRE_SNAP_THRESHOLD) {
+            wire.setIsStraight(true);
+            return c;
+        }
+        return x;
+    }
+
+    const v = obj.getPos();
+    // Snap to connections
+    for (const port of obj.getPorts()) {
+        const pos = port.getWorldTargetPos().sub(obj.getPos());
+        const wires = port.getWires();
+        for (const w of wires) {
+            // Get the port that isn't the current port
+            const port2 = (w.getP1() == port ? w.getP2() : w.getP1());
+            w.setIsStraight(false);
+            v.x = DoSnap(w, v.x + pos.x, port2.getWorldTargetPos().x) - pos.x;
+            v.y = DoSnap(w, v.y + pos.y, port2.getWorldTargetPos().y) - pos.y;
+        }
+    }
+    obj.setPos(v);
+}
 
 export class TranslateTool extends Tool {
     protected camera: Camera;
@@ -101,7 +127,10 @@ export class TranslateTool extends Tool {
                 newPositions;
 
         // Execute translate but don't save to group action since we do that in 'deactivate'
-        this.components.forEach((c, i) => new TranslateAction(c, shiftedPositions[i]).execute());
+        this.components.forEach((c, i) => c.setPos(shiftedPositions[i]));
+
+        // Snap at the end instead of one-by-one (fixes #417)
+        this.components.forEach(c => SnapPos(c));
 
         return true;
     }
